backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-groq>=0.1.5
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
google-generativeai>=0.8.0
bcrypt>=4.1.0
supabase>=2.4.0
pymupdf>=1.24.3
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
pydantic-settings>=2.2.1
httpx>=0.24.0